Move PartitionState into requests pkg.
