+++ clients/src/main/java/org/apache/kafka/common/requests/PartitionState.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package org.apache.kafka.common;
+package org.apache.kafka.common.requests;
 
 import java.util.Arrays;
 import java.util.List;
